Replace metre-conversion model with strict piece/meter product typing
Business rule changed: every product is now EITHER a piece carpet OR a
metre carpet, never a conversion between the two. This replaces the
meters_per_piece system from the previous session entirely.

- Product.unit_type: 'piece' | 'meter', fixed at creation (the update
  endpoint strips it from the request body so it can't change later --
  every downstream stock/sale/transfer record depends on it staying put).
- Stock/TransferItem/BranchSale/Purchase each keep independent quantity
  (pieces) and meter_quantity (metres) columns; exactly one is ever
  non-zero per row, decided by the product's unit_type. Nothing converts
  between them anymore.
- utils/meters.js rewritten: isMeterType/isPieceType, valueOf(stockRow,
  product) reads whichever field applies, applyStockDelta mutates the
  correct field and rejects a delta that would go negative.
- Purchase/Transfer/Sale controllers take a single `amount` field (native
  unit for that product) instead of the old quantity+meter_quantity pair
  with an optional override -- there's nothing to override now.
- Stock/dashboard endpoints aggregate piece and metre totals separately
  (never summed together, since they're different units) and expose
  unit_type on every stock/product row so the frontend knows which field
  to read.

Frontend: every purchase/transfer/sale form now shows exactly one amount
input, labelled and stepped for the selected product's unit (whole
pieces vs 0.01m). Products.js replaces the meters-per-piece field with a
type dropdown. Admin dashboard adds explicit "Dona mahsulotlari" / "Metr
mahsulotlari" totals per the spec, and its branch-debt drill-down now
shows the correct unit per transfer item instead of a piece count that
read 0 for metre products.

migrate.js repurposed for this schema change (its old meters-per-piece
backfill is obsolete); seed.js marks two sample carpets as unit_type
'meter' for local testing.

Verified end-to-end for both product types: purchase -> transfer
(including a same-warehouse-set with mixed piece/meter stock) -> sale,
with exact value/stock arithmetic and correct oversell rejection in both
the API and the form.

+++ backend/src/models/Stock.js
@@ -5,10 +5,10 @@ const Stock = sequelize.define('Stock', {
   id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
   warehouse_id: { type: DataTypes.INTEGER, allowNull: false },
   product_id: { type: DataTypes.INTEGER, allowNull: false },
-  // Metres are the source of truth for stock level and value.
-  // `quantity` is the piece equivalent, kept in sync as meters / meters_per_piece,
-  // so it is decimal (selling 20m off a 6m carpet leaves a fractional piece count).
-  quantity: { type: DataTypes.DECIMAL(12, 2), allowNull: false, defaultValue: 0 },
+  // Exactly one of these is ever non-zero for a given row, decided by the
+  // product's unit_type: 'piece' products use quantity, 'meter' products
+  // use meter_quantity. The two are never converted between each other.
+  quantity: { type: DataTypes.INTEGER, allowNull: false, defaultValue: 0 },
   meter_quantity: { type: DataTypes.DECIMAL(12, 2), allowNull: false, defaultValue: 0 },
 }, {
   tableName: 'stock',
